feat test database models
