+++ api/db/base.py
@@ -1,0 +1,9 @@
+from db.base_class import Base
+from db.models.user import User
+from db.models.apiKey import ApiKey
+from db.models.applicationType import ApplicationType
+from db.models.conference import Conference
+from db.models.council import Council
+from db.models.role import Role
+from db.models.sensible import Sensible
+from db.models.votingQuestion import VotingQuestion
