+++ api/db/models/apiKey.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, Date, ForeignKey
 from sqlalchemy.orm import relationship
-from db.models.base import Base
+from db.base import Base
 
 class ApiKey(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
@@ -8,8 +8,8 @@ class ApiKey(Base):
     note = Column(String(255), nullable=False)
     createdOn = Column(Date, nullable=False)
     validUntil = Column(Date, nullable=True)
-    conferenceId = Column(Integer, ForeignKey('conference.id'), nullable=False)
+    # conferenceId = Column(Integer, ForeignKey('conference.id'), nullable=False)
 
 
-    conference = relationship("conference", backref="apiKey")
+    # conference = relationship("conference", backref="apiKey")
