+++ api/db/models/votingQuestion.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Integer, String, Boolean, DateTime, JSON
-from db.models.base import Base
+from db.base import Base
 
 class VotingQuestion(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
